fix card create usage, rename controller file
